10/12/2022
Removed the usage of batch files. Now, we're going to use PowerShell scripts. These scripts are made to elevate to Administrator privileges to avoid any issues while running.

We now have scripts to install all the services, databases and to remove them as well. This is great for quick tests and removal. They will begin with, _. That way it's first in line and considered the most important ones to run first while setting up your server.