The automatic database installation now uses the C:/ drive as the default instead of the D:/ drive.
Previously, the backups were configured to use the D:/ drive, which caused errors for many users (majority) who had installed SQL Server to the default path. I have also included a disclaimer in the documentation, stating that if you have changed the default installation path for SQL, you will need to restore the databases manually instead of relying on the provided script.